// src/models/usersession.ts
export class UserSession {
    schoolName: String;
    schoolId: number;
    authToken: string;
    edu_dist_name: string;
    username: string;
    usertype: string;
    usertypeid: number;
    userId: number;
    schoolTypeId:number;
    highClass:number;
    lowClass:number;
    districtId:number;
    emisUsertype1:number;
    districtName:string;
    blockId:number;
    blockName:string;
    offKeyId:number;
  cateType: string;
  teacherType: any;
  edn_dist_name: string;
  teacherId : any;
  cate_id: any;
  catty_id:any;
  manage_id: any;
  edu_district_id: any;
  
  }